+++ crates/azure_types/src/policy_set_definitions.rs
@@ -2,11 +2,11 @@ use crate::prelude::PolicyDefinitionId;
 use crate::prelude::PolicySetDefinitionId;
 use crate::scopes::HasScope;
 use crate::scopes::Scope;
-use cloud_terrastodon_hcl_types::prelude::Sanitizable;
 use cloud_terrastodon_hcl_types::prelude::AzureRMResourceBlockKind;
 use cloud_terrastodon_hcl_types::prelude::HCLImportBlock;
 use cloud_terrastodon_hcl_types::prelude::HCLProviderReference;
 use cloud_terrastodon_hcl_types::prelude::ResourceBlockReference;
+use cloud_terrastodon_hcl_types::prelude::Sanitizable;
 use serde::Deserialize;
 use serde::Serialize;
 use serde_json::Value;
@@ -42,7 +42,7 @@ pub struct PolicySetDefinition {
     pub parameters: Option<HashMap<String, Value>>,
     pub policy_definitions: Option<Vec<PolicySetDefinitionPolicyDefinition>>,
     pub policy_definition_groups: Option<Vec<PolicySetDefinitionPolicyDefinitionGroup>>,
-    pub policy_type: String,
+    pub policy_type: String, // TODO: create enum
     pub version: String,
 }
 
@@ -70,13 +70,51 @@ impl std::fmt::Display for PolicySetDefinition {
     }
 }
 impl From<PolicySetDefinition> for HCLImportBlock {
-    fn from(policy_definition: PolicySetDefinition) -> Self {
+    fn from(policy_set_definition: PolicySetDefinition) -> Self {
         HCLImportBlock {
             provider: HCLProviderReference::Inherited,
-            id: policy_definition.id.expanded_form().to_string(),
+            id: policy_set_definition.id.expanded_form().to_string(),
             to: ResourceBlockReference::AzureRM {
                 kind: AzureRMResourceBlockKind::PolicySetDefinition,
-                name: policy_definition.name.sanitize(),
+                name: {
+                    let prefix = match policy_set_definition.id {
+                        PolicySetDefinitionId::Unscoped(_) => None,
+                        PolicySetDefinitionId::ManagementGroupScoped(
+                            management_group_scoped_policy_definition_id,
+                        ) => Some(
+                            management_group_scoped_policy_definition_id
+                                .management_group_id
+                                .short_form(),
+                        ),
+                        PolicySetDefinitionId::SubscriptionScoped(
+                            subscription_scoped_policy_definition_id,
+                        ) => Some(
+                            subscription_scoped_policy_definition_id
+                                .subscription_id
+                                .short_form(),
+                        ),
+                        PolicySetDefinitionId::ResourceGroupScoped(
+                            resource_group_scoped_policy_definition_id,
+                        ) => Some(
+                            resource_group_scoped_policy_definition_id
+                                .resource_group_id
+                                .short_form(),
+                        ),
+                        PolicySetDefinitionId::ResourceScoped(
+                            resource_scoped_policy_definition_id,
+                        ) => Some(
+                            resource_scoped_policy_definition_id
+                                .resource_id
+                                .short_form(),
+                        ),
+                    };
+
+                    match prefix {
+                        None => policy_set_definition.name,
+                        Some(prefix) => format!("{}_{}", prefix, policy_set_definition.name),
+                    }
+                    .sanitize()
+                },
             },
         }
     }
@@ -84,10 +122,12 @@ impl From<PolicySetDefinition> for HCLImportBlock {
 
 #[cfg(test)]
 mod tests {
-    use crate::scopes::{TryFromManagementGroupScoped, TryFromSubscriptionScoped};
+    use crate::scopes::TryFromManagementGroupScoped;
+    use crate::scopes::TryFromSubscriptionScoped;
 
     use super::*;
-    use eyre::{bail, Result};
+    use eyre::Result;
+    use eyre::bail;
 
     #[test]
     fn unscoped() -> Result<()> {
@@ -95,8 +135,8 @@ mod tests {
         let id = PolicySetDefinitionId::try_from_expanded(expanded)?;
         assert_eq!(id.expanded_form(), expanded);
         match &id {
-            PolicySetDefinitionId::Unscoped(_) => {},
-            x => bail!("Bad: {x:?}")
+            PolicySetDefinitionId::Unscoped(_) => {}
+            x => bail!("Bad: {x:?}"),
         }
         assert_eq!(id.short_form(), "my-policy-set-name");
         Ok(())
@@ -109,8 +149,8 @@ mod tests {
         assert_eq!(id, PolicySetDefinitionId::try_from_expanded(expanded)?);
         assert_eq!(id.expanded_form(), expanded);
         match &id {
-            PolicySetDefinitionId::ManagementGroupScoped(_) => {},
-            x => bail!("Bad: {x:?}")
+            PolicySetDefinitionId::ManagementGroupScoped(_) => {}
+            x => bail!("Bad: {x:?}"),
         }
         assert_eq!(id.short_form(), "my-policy-set-name");
         Ok(())
@@ -123,8 +163,8 @@ mod tests {
         assert_eq!(id, PolicySetDefinitionId::try_from_expanded(expanded)?);
         assert_eq!(id.expanded_form(), expanded);
         match &id {
-            PolicySetDefinitionId::SubscriptionScoped(_) => {},
-            x => bail!("Bad: {x:?}")
+            PolicySetDefinitionId::SubscriptionScoped(_) => {}
+            x => bail!("Bad: {x:?}"),
         }
         assert_eq!(id.short_form(), "my-policy-set-name");
         Ok(())