+++ crates/azure_types/src/query_response.rs
@@ -5,7 +5,6 @@ use serde::Deserializer;
 use serde::Serialize;
 use serde::de::Error;
 use serde_json::Value;
-use std::any::type_name;
 
 #[derive(Debug, Serialize)]
 pub struct ResourceGraphQueryResponse<T> {
@@ -92,7 +91,7 @@ where
         #[cfg(debug_assertions)]
         let record = serde_json::from_value(Value::Object(map.clone())).wrap_err(format!(
             "failed to deserialize entry {i} as {}, map={map:#?}",
-            type_name::<T>()
+            std::any::type_name:: <T>()
         ))?;
         #[cfg(not(debug_assertions))]
         let record = serde_json::from_value(Value::Object(map))