+++ crates/hcl/src/lib.rs
@@ -5,7 +5,7 @@ mod body_formatter;
 mod data_block_creation;
 mod data_lookup_holder;
 mod data_reference_patcher;
-mod default_attribute_removal_patcher;
+mod default_attribute_cleanup_patcher;
 mod import_builder;
 mod import_lookup_holder;
 mod imported_resource_reference_patcher;
@@ -19,14 +19,19 @@ mod user_id_reference_patcher;
 mod work_dir_lifecycle;
 mod writer;
 pub mod prelude {
+    pub use crate::azuredevops_git_repository_initialization_patcher::*;
     pub use crate::block_lister::*;
+    pub use crate::data_reference_patcher::*;
+    pub use crate::default_attribute_cleanup_patcher::*;
     pub use crate::import_builder::*;
+    pub use crate::imported_resource_reference_patcher::*;
     pub use crate::importer::*;
+    pub use crate::json_patcher::*;
     pub use crate::provider_manager::*;
     pub use crate::reflow::*;
+    pub use crate::terraform_block_extracter_patcher::*;
+    pub use crate::user_id_reference_patcher::*;
     pub use crate::work_dir_lifecycle::*;
     pub use crate::writer::*;
-    pub use crate::json_patcher::*;
-    
     pub use cloud_terrastodon_hcl_types::prelude::*;
 }